app/requirements.txt
Flask
hvac
mysql-connector-python
cryptography